fix warnings at compilation

+++ VTFLib/MemoryReader.cpp
@@ -80,7 +80,7 @@ vlUInt CMemoryReader::Seek(vlLong lOffset, vlUInt uiMode)
 		return 0;
 	}
 
-	switch(uiMode)
+	switch((int)uiMode)
 	{
 		case 0:
 			this->uiPointer = 0;
